+++ Model/Export/AbstractExport.php
@@ -40,11 +40,6 @@ class AbstractExport
      * @var null|\TurnTo\SocialCommerce\Logger\Monolog
      */
     protected $logger = null;
-    
-    /**
-     * @var \Magento\Framework\Encryption\EncryptorInterface|null
-     */
-    protected $encryptor = null;
 
     /**
      * @var \Magento\Framework\Intl\DateTimeFactory|null
@@ -87,7 +82,6 @@ class AbstractExport
      * @param \TurnTo\SocialCommerce\Helper\Config $config
      * @param \Magento\Catalog\Model\ResourceModel\Product\CollectionFactory $productCollectionFactory
      * @param \TurnTo\SocialCommerce\Logger\Monolog $logger
-     * @param \Magento\Framework\Encryption\EncryptorInterface $encryptor
      * @param \Magento\Framework\Intl\DateTimeFactory $dateTimeFactory
      * @param \Magento\Framework\Api\SearchCriteriaBuilder $searchCriteriaBuilder
      * @param \Magento\Framework\Api\FilterBuilder $filterBuilder
@@ -99,7 +93,6 @@ class AbstractExport
         \TurnTo\SocialCommerce\Helper\Config $config,
         \Magento\Catalog\Model\ResourceModel\Product\CollectionFactory $productCollectionFactory,
         \TurnTo\SocialCommerce\Logger\Monolog $logger,
-        \Magento\Framework\Encryption\EncryptorInterface $encryptor,
         \Magento\Framework\Intl\DateTimeFactory $dateTimeFactory,
         \Magento\Framework\Api\SearchCriteriaBuilder $searchCriteriaBuilder,
         \Magento\Framework\Api\FilterBuilder $filterBuilder,
@@ -110,7 +103,6 @@ class AbstractExport
         $this->config = $config;
         $this->productCollectionFactory = $productCollectionFactory;
         $this->logger = $logger;
-        $this->encryptor = $encryptor;
         $this->dateTimeFactory = $dateTimeFactory;
         $this->searchCriteriaBuilder = $searchCriteriaBuilder;
         $this->filterBuilder = $filterBuilder;
@@ -154,7 +146,7 @@ class AbstractExport
      * @param \Magento\Store\Api\Data\StoreInterface $store
      * @return \Magento\Catalog\Model\ResourceModel\Product\Collection
      */
-    protected function getProducts(\Magento\Store\Api\Data\StoreInterface $store)
+    public function getProducts(\Magento\Store\Api\Data\StoreInterface $store)
     {
         $collection = $this->productCollectionFactory->create()
             ->addAttributeToSelect('id')