+++ Model/Import/AbstractImport.php
@@ -33,11 +33,6 @@ class AbstractImport
     protected $productFactory = null;
 
     /**
-     * @var \Magento\Framework\Encryption\EncryptorInterface|null
-     */
-    protected $encryptor = null;
-
-    /**
      * @var \Magento\Catalog\Model\Indexer\Product\Eav\Processor|null
      */
     protected $productEavIndexProcessor = null;
@@ -48,7 +43,6 @@ class AbstractImport
      * @param \TurnTo\SocialCommerce\Helper\Config $config
      * @param \TurnTo\SocialCommerce\Logger\Monolog $logger
      * @param \Magento\Catalog\Model\ProductFactory $productFactory
-     * @param \Magento\Framework\Encryption\EncryptorInterface $encryptor
      * @param \Magento\Store\Model\StoreManagerInterface $storeManager
      * @param \Magento\Catalog\Model\Indexer\Product\Eav\Processor $productEavIndexProcessor
      */
@@ -56,14 +50,12 @@ class AbstractImport
         \TurnTo\SocialCommerce\Helper\Config $config,
         \TurnTo\SocialCommerce\Logger\Monolog $logger,
         \Magento\Catalog\Model\ProductFactory $productFactory,
-        \Magento\Framework\Encryption\EncryptorInterface $encryptor,
         \Magento\Store\Model\StoreManagerInterface $storeManager,
         \Magento\Catalog\Model\Indexer\Product\Eav\Processor $productEavIndexProcessor
     ) {
         $this->config = $config;
         $this->logger = $logger;
         $this->productFactory = $productFactory;
-        $this->encryptor = $encryptor;
         $this->storeManager = $storeManager;
         $this->productEavIndexProcessor = $productEavIndexProcessor;
     }