+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.1 (2017-10-04)
+* Add option have a "from" date when exporting historical orders ([#18](https://github.com/turnto/magento2-turnto-socialcommerce/issues/18))
+* Make it possible for a developer to group reviews for similar products by modifying g:item_group_id value in the catalog feed ([#17](https://github.com/turnto/magento2-turnto-socialcommerce/pull/17)) 
+* Fix catalog/order feed export issue ([#16](https://github.com/turnto/magento2-turnto-socialcommerce/issues/16))
+* Misc bug fixes
+
 ## 1.1.0 (2017-09-24)
 * Add SSO (Single Sign On) feature ([#14](https://github.com/turnto/magento2-turnto-socialcommerce/issues/14))
 * Add support for using child/variant SKUs instead of parent SKUs ([#8](https://github.com/turnto/magento2-turnto-socialcommerce/issues/8))